+++ NEWS.md
@@ -267,8 +267,7 @@ xkbcommon [1.14.0-beta2] – 2026-09-10
   - `struct xkb_events` (new):
     - `struct xkb_events_config`
     - `xkb_events::xkb_events_new()`
-    - `xkb_events::xkb_events_ref()`
-    - `xkb_events::xkb_events_unref()`
+    - `xkb_events::xkb_events_destroy()`
     - `xkb_events::xkb_events_next()`
   - `enum xkb_event_type` (new)
   - `struct xkb_event` (new):